resources/test/tests/idlharness/IdlInterface/test_exposed_attribute.html

<!DOCTYPE HTML>
<html>

<head>
  <title>IdlInterface Exposed attributes</title>
</head>

<body>
  <div id="log"></div>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/resources/WebIDLParser.js"></script>
  <script src="/resources/idlharness.js"></script>
  <script src="../helper.js"></script>
  <script>
    "use strict";

    test(function () {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
          [Exposed=(Window),Exposed=(Worker)]
          interface Document {};
      `);
      idlArray.assert_throws(
          'Multiple \'Exposed\' extended attributes on Document',
          i => i.test());
    }, 'should throw for multiple exposed attributes.');

    test(function () {
      const idlArray = new IdlArray();
      idlArray.add_idls(`
          [Exposed=Window]
          interface NodeList {};

          [Exposed=(Window,Worker)]
          partial interface NodeList {
            readonly attribute unsigned long length;
          };
        `);
      idlArray.assert_throws('Foo', i => i.test());
    }, 'should throw for non-subset exposed attribute on partial.');

    // test(function () {
    //   const idlArray = new IdlArray();
    //   idlArray.add_idls(`
    //     [Exposed=Window]
    //     interface NodeList {};

    //     [Exposed=Window]
    //     partial interface NodeList {
    //       readonly attribute unsigned long length;
    //     };
    //   `);
    //   idlArray.test();
    // }, 'should not throw for exposed attribute on partial.');
  </script>
</body>

</html>
